+++ src/dgisim/effect/enums.py
@@ -13,48 +13,48 @@ __all__ = [
 
 
 class Zone(Enum):
-    CHARACTERS = "Characters"
-    SUMMONS = "Summons"
-    SUPPORTS = "Supports"
-    HIDDEN_STATUSES = "Hidden-Statuses"
-    COMBAT_STATUSES = "Combat-Statuses"
+    CHARACTERS = 1
+    SUMMONS = 2
+    SUPPORTS = 3
+    HIDDEN_STATUSES = 4
+    COMBAT_STATUSES = 5
 
 
 class TriggeringSignal(Enum):
     #: triggers prepare skill statuses
-    ACT_PRE_SKILL = "ActPreSkill"
+    ACT_PRE_SKILL = 1
     #: triggers when "after a character uses ..." (any skill)
-    COMBAT_ACTION = "CombatAction"
+    COMBAT_ACTION = 2
     #: triggers when "after a ... character is defeated"
-    DEATH_EVENT = "DeathEvent"
+    DEATH_EVENT = 3
     #: triggers when "at the end phase, ..."
-    END_ROUND_CHECK_OUT = "EndRoundCheckOut"
+    END_ROUND_CHECK_OUT = 4
     #: triggers when a fast action is performed, unused signal.
-    FAST_ACTION = "FastAction"
+    FAST_ACTION = 5
     #: triggers when "when the battle begins"
-    GAME_START = "GameStart"
+    GAME_START = 6
     #: triggers when "when your opponent declare the end of their round"
-    OPPO_DECLARE_END_ROUND = "OppoDeclareEndRound"
+    OPPO_DECLARE_END_ROUND = 7
     #: triggers when "when ... play a ... card"
-    POST_CARD = "PostCard"
+    POST_CARD = 8
     #: triggers when "when ... character takes DMG"
-    POST_DMG = "PostDmg"
+    POST_DMG = 9
     #: triggers when "after ... character takes Elemental Reaction DMG"
-    POST_REACTION = "PostReaction"
+    POST_REACTION = 10
     #: triggers when "before ... choose their action"
-    PRE_ACTION = "PreAction"
+    PRE_ACTION = 11
     #: triggers when "End Phase"
-    ROUND_END = "RoundEnd"  # remove frozen etc.
+    ROUND_END = 12  # remove frozen etc.
     #: triggers when "when the Action Phase starts"
-    ROUND_START = "RoundStart"
+    ROUND_START = 13
     #: triggers when "when you declare the end of your round"
-    SELF_DECLARE_END_ROUND = "SelfDeclareEndRound"
+    SELF_DECLARE_END_ROUND = 14
     #: ABOUT TO BE DEPRECATED: triggers when player 1 switch a character
-    SWAP_EVENT_1 = "SwapEvent1"  # P1's swap; TODO: make it relative
+    SWAP_EVENT_1 = 15  # P1's swap; TODO: make it relative
     #: ABOUT TO BE DEPRECATED: triggers when player 2 switch a character
-    SWAP_EVENT_2 = "SwapEvent2"  # P2's swap; TODO: make it relative
+    SWAP_EVENT_2 = 16  # P2's swap; TODO: make it relative
     #: triggers when "when the character ... would be defeated"
-    TRIGGER_REVIVAL = "TriggerRevival"
+    TRIGGER_REVIVAL = 17
 
     @classmethod
     def swap_event(cls, pid: Pid) -> TriggeringSignal:
@@ -71,10 +71,10 @@ class TriggeringSignal(Enum):
 
 
 class DynamicCharacterTarget(Enum):
-    SELF_SELF = 0
-    SELF_ACTIVE = 1
-    SELF_OFF_FIELD = 2
-    SELF_ALL = 3
-    SELF_ABS = 4
-    OPPO_ACTIVE = 5
-    OPPO_OFF_FIELD = 6
+    SELF_SELF = 1
+    SELF_ACTIVE = 2
+    SELF_OFF_FIELD = 3
+    SELF_ALL = 4
+    SELF_ABS = 5
+    OPPO_ACTIVE = 6
+    OPPO_OFF_FIELD = 7